Change GetArea return type from double to float
Updated the GetArea method to return a float instead of a double.
The area calculation result is now explicitly cast to float before
being returned.

+++ C#/Day 6/Lab6/Lab6.Demo2/clsCircle.cs
@@ -20,9 +20,9 @@
             return true;
         }
 
-        public double GetArea()
+        public float GetArea()
         {
-            return by * Math.Pow(Diminsion1, 2);
+            return (float)(by * Math.Pow(Diminsion1, 2));
         }
     }
 }
